+++ src/components/GestionSolicitudes.tsx
@@ -1,10 +1,11 @@
 import React, { useEffect, useState } from "react";
-import FiltroGS from "../../components/FiltroGestionSolicitudes";
-import Pagination from "../../components/Pagination";
+import FiltroGS from "./FiltroGestionSolicitudes";
+import Pagination from "./Pagination";
+import { NavLink } from "react-router-dom";
 
-const ActividadesSolicitadasEstudiantes: React.FC = () => {
+const GestionSolicitudes: React.FC = () => {
     useEffect(() => {
-        document.title = "Estudiantes - UNAH CUROC";
+        document.title = "Activdades Solicitadas - UNAH COPAN";
     }, []);
 
     
@@ -96,13 +97,23 @@ const ActividadesSolicitadasEstudiantes: React.FC = () => {
                                         <span className="inline-block w-1/3 md:hidden font-bold mr-4">Estado:</span>{item.estado}
                                     </td>
                                     <td className="p-1 md:border text-center md:border-gray-500 block md:table-cell relative">
-                                        <a href="#" className="flex justify-center items-center font-bold group"  title="Unirse">
-                                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width={32} height={32} color={"#000000"} fill={"none"} className="group-hover:text-blue-500 hidden md:block">
+                                    <NavLink
+                                        to={
+                                            location.pathname.includes('dashboard-coordinador')
+                                                ? "#"
+                                                : location.pathname.includes('dashboard-estudiante')
+                                                    ? "#"
+                                                    : "/dashboard-voae/gestion-solicitud"
+                                        }
+                                        className="flex justify-center items-center font-bold group"
+                                    >
+                                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width={32} height={32} color={"#000000"} fill={"none"} className="group-hover:text-blue-500 hidden md:block">
                                                 <path d="M21.544 11.045C21.848 11.4713 22 11.6845 22 12C22 12.3155 21.848 12.5287 21.544 12.955C20.1779 14.8706 16.6892 19 12 19C7.31078 19 3.8221 14.8706 2.45604 12.955C2.15201 12.5287 2 12.3155 2 12C2 11.6845 2.15201 11.4713 2.45604 11.045C3.8221 9.12944 7.31078 5 12 5C16.6892 5 20.1779 9.12944 21.544 11.045Z" stroke="currentColor" strokeWidth="1.5" />
                                                 <path d="M15 12C15 10.3431 13.6569 9 12 9C10.3431 9 9 10.3431 9 12C9 13.6569 10.3431 15 12 15C13.6569 15 15 13.6569 15 12Z" stroke="currentColor" strokeWidth="1.5" />
                                             </svg>
                                             <span className="hover:text-blue-500 group-hover:text-blue-500 md:hidden">Ver detalles</span>
-                                        </a>
+                                    </NavLink>
+                                       
                                     </td>
                                 </tr>
                             ))}
@@ -116,5 +127,5 @@ const ActividadesSolicitadasEstudiantes: React.FC = () => {
     );
 };
 
-export default ActividadesSolicitadasEstudiantes;
+export default GestionSolicitudes;
