+++ src/layouts/Layout.tsx
@@ -1,7 +1,7 @@
 
 import { Outlet } from 'react-router-dom';
-import Header from '../Header';
-import Footer from '../Footer';
+import Header from '../components/Header';
+import Footer from '../components/Footer';
 
 export default function Layout(){
   return (
